QueueFamilyIndices moved to new file
