+++ src/queue_family_indices.rs
@@ -1,0 +1,49 @@
+use vulkanalia::prelude::v1_0::*;
+use vulkanalia::vk::KhrSurfaceExtension;
+
+use anyhow::{anyhow, Result};
+use thiserror::Error;
+
+#[derive(Debug, Error)]
+#[error("Missing {0}.")]
+pub struct SuitabilityError(pub &'static str);
+
+#[derive(Copy, Clone, Debug)]
+pub struct QueueFamilyIndices {
+    pub graphics: u32,
+    pub present: u32,
+}
+
+impl QueueFamilyIndices {
+    pub unsafe fn get(
+        instance: &Instance,
+        surface: &vk::SurfaceKHR,
+        physical_device: vk::PhysicalDevice,
+    ) -> Result<Self> {
+        let properties = instance
+            .get_physical_device_queue_family_properties(physical_device);
+
+        let graphics = properties
+            .iter()
+            .position(|p| p.queue_flags.contains(vk::QueueFlags::GRAPHICS))
+            .map(|i| i as u32);
+
+        let mut present = None;
+        for (index, properties) in properties.iter().enumerate() {
+            if instance.get_physical_device_surface_support_khr(
+                physical_device,
+                index as u32,
+                *surface,
+            )? {
+                present = Some(index as u32);
+                break;
+            }
+        }
+
+        if let (Some(graphics), Some(present)) = (graphics, present) {
+            Ok(Self { graphics, present })
+        } else {
+            Err(anyhow!(SuitabilityError("Missing required queue families.")))
+        }
+    }
+}
